Prevent issue with empty $disallowed_id removing

+++ admin/admin_disallow.php
@@ -38,15 +38,17 @@ if (isset($_POST['add_name'])) {
 } elseif (isset($_POST['delete_name'])) {
     $disallowed_id = isset($_POST['disallowed_id']) ? (int)$_POST['disallowed_id'] : (int)$_GET['disallowed_id'];
 
-    $sql = 'DELETE FROM ' . BB_DISALLOW . " WHERE disallow_id = $disallowed_id";
-    $result = DB()->sql_query($sql);
-    if (!$result) {
-        bb_die('Could not removed disallowed user');
+    if (!empty($disallowed_id)) {
+        $sql = 'DELETE FROM ' . BB_DISALLOW . " WHERE disallow_id = $disallowed_id";
+        $result = DB()->sql_query($sql);
+        if (!$result) {
+            bb_die('Could not removed disallowed user');
+        }
+
+        $message .= $lang['DISALLOWED_DELETED'] . '<br /><br />' . sprintf($lang['CLICK_RETURN_DISALLOWADMIN'], '<a href="admin_disallow.php">', '</a>') . '<br /><br />' . sprintf($lang['CLICK_RETURN_ADMIN_INDEX'], '<a href="index.php?pane=right">', '</a>');
+
+        bb_die($message);
     }
-
-    $message .= $lang['DISALLOWED_DELETED'] . '<br /><br />' . sprintf($lang['CLICK_RETURN_DISALLOWADMIN'], '<a href="admin_disallow.php">', '</a>') . '<br /><br />' . sprintf($lang['CLICK_RETURN_ADMIN_INDEX'], '<a href="index.php?pane=right">', '</a>');
-
-    bb_die($message);
 }
 
 /**
